Split min-curve into unit aware/unaware functions
Separate out the unit aware and dependent functionality from
min_curve_method into a separate function.

The main goal is to provide a more layered interface, with escape
hatches if necessary. If the survey is odd, and with mixed units, the
min_curve_method would often be wrong. By separating the pure
mathematical operations into a separate functions, it can still be
useful with the assumption that care is taken to make the units
consistent.