Interpolation as test strategy for curve equality
Use interpolation onto a new interval as test strategy for comparing
reference curves with those output by the interpolate functions. DTW had
the immediate drawback that it was quite costly to run, and the distance
value was hard to make sense of.

Since we *know* the output of the interpolate-functions to be equivalent
(differently sampled and some numerical noise), a faster and simpler
strategy is to simply interpolate the reference curve and the computed
curve onto a new, now identical, sample interval and comparing them
numerically.

As a consequence, the dependency on dtw and similaritymeasures goes
away.