Remove units from minimum curvature method
Following the new API, remove the units from the minimum_curvature
method. It is now the user's responsibility to set correct units
using wp.unit_convert.

Most commonly units for md should be meters.

Test cases modified to match this new API.

norm_opt is replaced by course_length which is more correct drilling
nomenclature.
course_length is set to the default value of 30 for md in meters.
It can be set to any value but will most likely be set to 100 for
md in feet.

Other values are not recommended.