Check geometry.tangent domain properties
The tangent of inclination and azimuth points has some properties
that must be respected. These tests verifies two such properties:
for an inclination value of zero, the tangent is always vertical,
and irrespective of the input values of inclination and azimuth,
the tangent outputs (vd, northing and easting) are always within
the allowed range.