Check geometry.tangent behaviour with systematic vectors
The tangent of inclination and azimuth points has some
properties that must be respected. These tests verify
that for a random vector in each of the eight octants
of a sphere, the resulting tangents match the expected
values.

In order to test this behaviour without using our own
code, we passed eight different vectors to
www.wolframalpha.com and used the spherical coordinates
generated as inputs to `tangent()`, and the polar
coordinates generated as outputs to verify our code
against.