Add deviation and position_log with resample
The new interface for deviation and position logs abstract away the
array details, and help callers focus on the mapping between log types.
There's often a lot of non-trivial book keeping involved in converting
between formats, so an object that maintains this state is very helpful.

The usability is a net win, although only min-curve support is added for
now. The position log -> deviation is missing and will always raise
NotImplemented for now, as it is probably method dependent, and seems
quite hard to do accurately.

What is really new here is the position_log.resample method. This
immediately deprecates all previous interpolate_* methods, both as it
conveys the actual transformation (the position log sampled on a new
measured depth), and because it is a lot more correct for min-curve
because of its use of spherical interpolation.

The idea and implementation is largely inspired by David M. Baker's
arc2chord.