Add vector functions to geometry module
Add a small set of functions for vectors to the geometry module, and
pull equivalent helpers out of position_log.py.

With the development of the deviation() function it has been
demonstrated that a set of functions for computing on vectors are
necessary, and already informally implemented in helpers for slerp and
more.

Most of the implementations are just taken straight out of the text book
definitions, with some additional tuning for floating point shenanigans.