Correct the resample() min-curve position log
The implementation of the resample function for the min-curve position
log was a buggy mess, and had several errors. The fundamental approach
has been cleaned up, and the algorithm more documented and understood.

The function signature is unchanged, but the concept of a resampled_md
member is gone, as this is not necessary to compute anything, and can
now be derived from a properly resampled log.

Getting this right was a lot of work - the fundamental approach is
inspired by David Baker's arc2chord [1], but redesigned to work with the
interface and flow of wellpathpy. Notably, the interpolation is no
longer done on a unit circle from which new inc/azimuths are computed,
but rather solved directly in cartesian.

The largest issue was getting the geometry right, and the final pieces
was put in place with superb help from Jens G. Magnus and Alena
Chaikouskaya.

There is a minor tuning to the internal spherical_interpolate function,
namely the omega/angle as a parameter, since computing the right angle
needs more information than just the two points. It could possibly be
derived from the points, but making it explicit makes for a nice flow.

[1] https://github.com/dmbaker/wellpathpy/commit/6277996ef4dc4dcec7980d918d19cf0151f4fa64