Use numpy instead of pandas to read CSV deviations
In order to decrease denpendancies, we are dropping pandas in favor
of numpy to load CSV files. As a result there is an API change in
read_csv that is a *breaking* change; indeed we no longer parse the
columns names in a header row and we make some assumptions about
the input file format and content. Namely (1) that input files
contain a header row and (2) that the columns are ordered with
`md`,`ìnc`,`azi`.