Update tests for new read_csv API
Swapping pandas for numpy to read CSV files has some impacts on
the tests that we run: we can no longer test for mixed column order,
we must still check to see if too few columns are passed, but can
no longer check their keys as the header row -if present- is not
parsed. New lines at the end of the files are no longer accepted.