Don't do trig inline in mincurve
Use the direction_vector computed vectors rather than inline
trigonometry to compute the position log. These are computed anyway in
order to use angle_between for the dogleg, so it makes sense to reuse
them. On top of doing less work this improves the clarity of the
function significantly, at the cost of move it one step away from the
textbook formulae.