Snap zero-angle ratio factor correctly
When the dogleg angle is zero (no change between two survey stations)
there is no dampening effect and the ratio factor should be one 1, but
the divide-by-zero must be handled. This already happens in practice for
very small angles [1] but the implementation was buggy and applied the 1
to the dogleg, not the rf.

[1]
>>> def rf(x): return (2 * np.tan(x/2))/x
>>> rf(1e-10)
1.0